pyproject.toml
[project]
name = "bmad-mcp"
version = "0.1.0"
description = "MCP server for BMAD Autopilot - orchestrates story creation, development, and code review"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
bmad-mcp = "bmad_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bmad_mcp"]
